Added Options to Change Colors and Other Minor Stuff
Idk why I did this, I didn't want to invest this much time in this project but whatever, now it's there and it looks good. Also added a label for the current generation in the visualization and fixed a few small things (f.ex.: The entities are now not glitching into the obstacles anymore, or at least not in a noticable manner).

Wow, what a nice commit message this is.